+++ Changelog.md
@@ -7,6 +7,7 @@ for planned releases.
 
 * `krillc issues` fails with `Error: Unknown API method` (#248)
 * `krillc parents` help text refers incorrectly to publisher request instead of child request (#251)
+* Normalize request/response `krillc help` texts (#252)
 
 ## 0.6.2 Release 'That was even faster!'